+++ lib/Braintree/Error/Codes.php
@@ -634,13 +634,17 @@ class Codes
     const TRANSACTION_TAX_AMOUNT_IS_TOO_LARGE                                         = '81536';
 
     const TRANSACTION_THREE_D_SECURE_AUTHENTICATION_FAILED                            = '81571';
+    const TRANSACTION_THREE_D_SECURE_AUTHENTICATION_RESPONSE_IS_INVALID               = '915120';
+    const TRANSACTION_THREE_D_SECURE_CAVV_ALGORITHM_IS_INVALID                        = '915122';
+    const TRANSACTION_THREE_D_SECURE_CAVV_IS_REQUIRED                                 = '915116';
+    const TRANSACTION_THREE_D_SECURE_DIRECTORY_RESPONSE_IS_INVALID                    = '915121';
+    const TRANSACTION_THREE_D_SECURE_ECI_FLAG_IS_INVALID                              = '915114';
+    const TRANSACTION_THREE_D_SECURE_ECI_FLAG_IS_REQUIRED                             = '915113';
+    const TRANSACTION_THREE_D_SECURE_MERCHANT_ACCOUNT_DOES_NOT_SUPPORT_CARD_TYPE      = '915131';
+    const TRANSACTION_THREE_D_SECURE_THREE_D_SECURE_VERSION_IS_INVALID                = '915119';
     const TRANSACTION_THREE_D_SECURE_TOKEN_IS_INVALID                                 = '91568';
     const TRANSACTION_THREE_D_SECURE_TRANSACTION_DATA_DOESNT_MATCH_VERIFY             = '91570';
-    const TRANSACTION_THREE_D_SECURE_ECI_FLAG_IS_REQUIRED                             = '915113';
-    const TRANSACTION_THREE_D_SECURE_CAVV_IS_REQUIRED                                 = '915116';
     const TRANSACTION_THREE_D_SECURE_XID_IS_REQUIRED                                  = '915115';
-    const TRANSACTION_THREE_D_SECURE_ECI_FLAG_IS_INVALID                              = '915114';
-    const TRANSACTION_THREE_D_SECURE_MERCHANT_ACCOUNT_DOES_NOT_SUPPORT_CARD_TYPE      = '915131';
     const TRANSACTION_TYPE_IS_INVALID                                                 = '91523';
     const TRANSACTION_TYPE_IS_REQUIRED                                                = '91524';
     const TRANSACTION_UNSUPPORTED_VOICE_AUTHORIZATION                                 = '91539';